Adressing PR feedback. Don't create new api version. Just use previous version. Make new entity relation Restrict instead of Cascade.
