--- src/Maestro/Maestro.Web/Api/v2019_01_10/Models/UpdateFrequency.cs
@@ -1,13 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-namespace Maestro.Web.Api.v2019_01_10.Models
-{
-    public enum UpdateFrequency
-    {
-        None = 0,
-        EveryDay,
-        EveryBuild
-    }
-}
